Removed unnecessary use of DelayedInit.
Realised that you don't really need DelayedInit at all in most cases - it's not necessary to run wiring after the object's constructor has run. It is fine to run it at any point, so we can just run it as part of the trait's initialiser. No magic code needed, other than the regular magic of Scala.